ci: add smoke test
Add a simple smoke test that checks whether the USB controller pops up
in the VMs PCI config space.

 As the xhci driver currently hangs for quite a long time, we only
check whether the kernel sees it during boot up. This makes the test
succeed in a couple of seconds. If we wait for VM userspace to come
up, this currently takes more than a minute.

We can refactor the test later when the xhci driver attaches
cleanly.