app: make dma_bus modifiable
Making Bus modifiable turns out to be difficult without making it
pretty ugly in the process. Instead I opted to layer some ArcSwap
magic on top of it. Modifications of the memory bus are super rare, so
we can afford rebuilding the underlying memory bus when a memory
region is added.