tests: inject an interrupt on start
Add a systemd user unit to dump diagnostic information after the
automatic login of the nixos user. Dump the interrupt counts for now and
check whether the xhci_hcd MSI-X interrupt fired once.

We inject a bogus interrupt when the controller run bit is toggled on.
The guest driver does not seem to complain about this, but we can see
the interrupt arrived in the guest.

Signed-off-by: Stefan Nürnberger <stefan.nuernberger@cyberus-technology.de>