smoke-test: load kvm module at second stage of boot
On my system, the smoke-test consistently (five attempts) failed,
because the systemd cloud-hypervisor.service failed ("Failed to open
hypervisor interface"). Manually restarting the service in an
interactive test run worked, so we assume there is a race condition
between the loading of KVM and the execution of the
cloud-hypervisor.service.
This commit ensures that KVM is loaded during the second stage of the
boot process and thus reliably available for the start of
cloud-hypervisor.service.