xhci_backend: provide correct DMA bus to XhciController
The XhciBackend and its associated XhciController should share the same
DMA bus. The backend performs the DMA map and unmap requests of the VFIO
client and the controller performs DMA reads and writes to the mapped
memory.
Previously, the backend and the controller received separate buses. This
behavior is a bug. This commit ensures that both components receive the
same DMA bus.