lib: fold into main application
The library was a convenient way to get the device emulation code over
from the old Particle code base. Now that we actually started to use
this code and added more modules in the main application, there is no
good distinction anymore between what's in the library and what's in
the main application.

So remove the library and treat the device module like all the other
modules.