xhci: use constants for doorbell addresses
The spec does not define the doorbell register offsets, and so far we
placed them on their own page (0x2000), but had no hard decision about
the topic. With more interaction, we should stabilize the decision, so
this commits defines the address offsets for the doorbell registers as
constants and uses them throughout the xhci controller.
The current plan is to support only one device, so we only need two
doorbells--one for the controller (command ring notification) and one
for the device (transfer ring notification). This commit chooses 0x2004
for the device doorbell, which means that the response to a Enable Slot
Command has to instruct the driver to use Device Slot 1.