rings: avoid recursion in command ring access
Sometimes, we cannot just return the current TRB, because the TRB is not
a command but a Link TRB, which tells our ring implementation that the
dequeue pointer has to jump to another (or the start of the same) ring
memory segment, and perhaps the cycle state flips.
This commit removes the recursion that we previously used to access the
first TRB in the new ring segment.