rings: EventRing owns a BusDeviceRef
Instead of passing a BusDeviceRef every time the EventRing needs one
(e.g., to enqueue an event), we initialize the EventRing with its own
reference.
This change makes it a little bit more unclear where DMA requests happen
when looking at the xhci controller code, but the gained convenience is
worth it.