rings: CommandRing owns a BusDeviceRef
Instead of passing a BusDeviceRef every time the CommandRing needs one
(to dequeue a command), we initialize the CommandRing with its own
reference.
This change makes it a little bit more unclear where DMA requests happen
when looking at the xhci controller code, but the gained convenience is
worth it.