xhci: acknowledge DisableSlot and StopEndpoint commands
The driver is unhappy after requesting the device descriptor on the
transfer rings six times and we do not provide any data in response. We
handle the DisableSlot and StopEndpoint commands by just acknowleding
them for now, so that we do not crash during the smoke-test. Later, more
handling of the two commands will most likely be necessary.