trb: use RawTrbBuffer (alias for [u8; 16]) to represent raw TRB bytes
So far, the `try_from` and `parse` methods of the `CommandTrb` and
`TransferTrb` expected byte slices. Consequently, size checks in the
`try_from` methods and assumptions about the size in the `parse` methods
were necessary.

This commit replaces the byte slices with [u8; 16] (all TRBs are 16
bytes large). The new type alias `RawTrbBuffer` for [u8; 16] makes even
clearer what `try_from`/`parse` expect, and also improves the
comprehension of the return value of the `to_bytes` method on the
`EventTrb`.