trb: refactor the CommandTrb type
Previously, we needed to pass the complex type `(address,
Result<CommandTrb, TrbParseError>)` around (e.g., return from
`next_command_trb`), because the `CommandTrb` only represented the data
encoded in the TRB. However, the address is also important, so that we
can send a `CommandCompletionEvent`, which needs to point to the Command
TRB.

Now, `CommandTrb` is called `CommandTrbVariant` and a field of the new
struct `CommandTrb`. `address` is the second field in `CommandTrb`.
Further, the error is now represented by
`CommandTrbVariant::Unrecognized`.

The changes also required adaptation of the related functions in the
rings and xhci controller code.