trb: refactor the TransferTrb type
Previously, we needed to pass the complex type `(address,
Result<TransferTrb, TrbParseError>)` around (e.g., return from
`next_transfer_trb`), because the `TransferTrb` only represented the
data encoded in the TRB. However, the address is also important, so that
we can send a `TransferEvent`, which needs to point to the Transfer TRB.

Now, the enum `TransferTrb` is called `TransferTrbVariant` and a field
of the new struct `TransferTrb`. `address` is the second field in
`TransferTrb`. Further, the error is now represented by
`TransferTrbVariant::Unrecognized`.

The changes also required adaptation of the functions that retrieve
transfer TRBs and requests in the
rings code.