usbrequest: refactor the UsbRequest type
There are two changes:

UsbRequest now also includes the address of the Status Stage TRB that
concluded this request. This address has to be placed into the
TransferEvent that marks the completion of the transfer. Previously, a
tuple of (address, UsbRequest) was passed around.

The constructor have been removed, as they provide little value and can
even lead to easier mistakes/worse readability.

The usage of UsbRequest in the rings and the xhci controller also
required small adaptations.