xhci: dequeue OUT transfer ring TRBs until empty
So far, we only dequeued a single TRB from a transfer ring when the
device doorbell was rang for the associated endpoint. If multiple TRBs
were present (apparently has not happened, but the situation is very
possible), we would only process the first. If no TRB is present, we
panic (that should be fine, I don't think the driver should ring the
doorbell when there is nothing for us to do).

This commit introduces a loop that dequeues and processes TRBs from an
OUT endpoint until the ring is empty. This changes matches the OUT
handling with the IN handling (where we have already observed that
multi-TRB processing is necessary).