xhci/constants: expose second port as USB2 port
so far, we reported that we only support a single port, which supported
USB 3.2. This port supports 3.0, 3.1, and 3.2 device, but does not
support USB 2. Further, the kernel is confused by us reporting a single
USB 3 port—usually, an USB-3-capable physical port would report both a
USB 2 as well as a USB 3 port (the kernel assumes there are two ports
both they both map to our single port).

This commit adds a second port and configures the port's protocol to USB
2. For now, we indicate with the associated PORTSC that no device is
connected. The functionality of the existing USB 3 port is unaffected.