real device: pass endpoint id to endpoint transfers
so far, we only assumed a single endpoint per direction (IN/OUT), so
there was no need to pass the endpoint along with requests. This
assumption will no longer be true with the following commits, so here we
change the trait functions of the real device to pass the endpoint id
along with transfer requests.