nusb: support enabling of any bulk endpoint
so far, we only supported endpoint 3 (EP1 IN) and 4 (EP2 OUT), because
these two were used by all USB 3 flashdrives that we tested. We
implemented more generic handling of the endpoints of a nusb device now,
because we already observed a USB 2 flashdrive using endpoints 2 (EP1
OUT) and 5 (EP2 IN) instead.