real device: device speed query
When we want to passthrough both USB 2 and 3 devices. It is important
that we indicate the correct USB speed to the guest driver, otherwise
the driver might later be unhappy when max_packet_size from the device
descriptor and the reported speed are an implausible combination. We
need to know the actual device speed to correctly indicate the same
information to the guest.