xhci: set portsc register on device attach
so far, we always set the USB 3 portsc register to indicate that there
is a superspeed device connected. Now, initially the xhci controller
reports that his usb 3 and usb 2 ports are powered but there are no
devices connected. When a device is attached (currently can only happen
very shortly after initialization), only then we indicate through portsc
that a device is available.

Further, we now check the speed of the device and determine whether to
advertise on the usb 3 or usb 2 port and which linkspeed to report in
portsc; instead of always indicating that the device is on the usb 3
port.