xhci: allow BSR bit in address device command
with the USB3 flashdrive, the BSR bit was always set. The panic (should
have been a todo) should make it obvious when we encounter an address
device command that has this bit set, so we can further analyze the
situation and requirements. The time has come now with USB 2, which
first issue Address Device with BSR set and later with BSR unset.
Because the flag controls USB address assigning, which we don't need
without a physical bus, there should be nothing special to handle (BSR
says "don't assign address", which we already don't do).