device-slots: report endpoint type of enabled endpoints
So far, we assumed that everything except the control endpoint EP0 is a
bulk endpoint. We want to support InterruptIn endpoints soon, so we need
to be able to distinguish the configured endpoint type, so we can access
the endpoints on the real device correctly. This commit implements the
functionality to extract the endpoint-type information from the input
context in the Configure Endpoint Command.