xhci_backend: reset device at launch
So far, we had the issue that passing through the same device multiple
times in succession let to the passthrough working every second time.
Each run after a successful run would find the device in an active state
where the device responds to the initial requests from the fresh VM with
stall errors. As a result of the errors, usbvfiod crashes (currently
intended design) and the device somehow changes state so that on the
next run, it responds normally to the inital requests from the fresh VM.

Now, we let nusb reset the device before we do anything with
it---passthrough now works every time.