realdevice: pass endpoint type to enable_endpoint
The endpoint-type information is relevant for enabling the correct
endpoints in the real device. This commit makes the endpoint type
available in the enable_endpoint function, but does not make use of it
yet (the enable_endpoint function will receive great rework to make the
existing bulk endpoint handling asynchronous, making changes to handle
interrupt endpoints does not make sense yet).