nusb: claim all interfaces
so far, we tested passthrough with only a single interface (usb sticks
and mouse), where our approach of claiming interface 0 and getting
access to the endpoints through this interface worked.
When testing passthrough with a logitech keyboard, which had two
interface with one interrupt IN endpoint each (one is for generic HID
keyboard input, the other not sure, perhaps vendor-specific), the guest
driver to enable both endpoints, but nusb stated that one of the
endpoints did not exist (because it belongs to another interface).

The xhci controller is not aware of which interfaces exist on the device
and which endpoint belong to which interface, but nusb implements these
checks to help userspace-driver developers catch errors in their
endpoint-opening code more easily.

This commit prepares for access to all endpoints: During initialization,
we claim all interfaces. However, the `enable_endpoint` code continues
to always use interface 0--this is subject to change in the follow-up
commit.