nusb: find correct interface for endpoint
This commit changes the behavior of `enable_endpoint` from always trying
to open an endpoint on interface 0 to opening the endpoint on the
interface it belongs to. We find the correct interface by going through
all interface and all endpoints that belong to them, until we find the
desired endpoint listed.