nusb: support control out transfers with data
We did not support such control out transfers with data yet, because we
did not need them. Control requests with in IN with data (e.g.,
requesting a descriptor) or Control out without data (e.g., setting
configuration).

When experimenting with interrupt endpoints and trying to passthrough a
keyboard, I encountered control out requests with data. The keyboard was
using SET_REPORT control transfers with 1 data byte, probably to set the
status of capslock/numlock LEDs.

This commit extends the handler for control out to also process requests
with data correctly.