nusb: allow all transfer lengths
nusb has the limitation that we can only send transfer IN requests when
the transfer length is a multiple of the maximum packet size of the
endpoint. So far, we only encountered two scenarios:
- transfer length smaller than the max packet size, e.g., 13 byte for
SCSI CSW messages
- transfer length as a multiple of the max packet size, typically for
sending USB storage content
Because we did not encounter the third case yet (transfer length larger
than max packet size, but not a multiple), we chose to panic when this
scenario occurs and investigate.

@Noi0103 encountered this panic when trying to create a partition on a
passthroughed USB stick (max packet size == 1024, transfer length ==
2560 == 1024 * 2.5). The only reasonable handling is to use the
next-largest multiple of max_packet_size, which this PR adds. The
partition creation works with this change applied.