nusb: implement worker thread for bulk transfer IN handling
This function takes the logic from XhciController::check_in_endpoint and
NusbDeviceWrapper::transfer_in and combines it with the ability to sleep
when no more TRBs are available, as well as wake-up functionality
(intended to by used by the doorbell).

The function is intended to run on a separate thread, but the
integration requires a lot of changed lines in the existing code, so it
is split into a separate commit.