nusb: get rid of obsolete EndpointWrapper
Previously, we stored the endpoints themselves, and had to wrap them
into an enum to store them in an array. Now, the endpoints are stored in
the worker threads, and we just store the mpsc::Sender that can wake up
the worker threads. No enum is needed, we can directly store the senders
in array.