nusb: respect interrupt_on_completion flag
so far, we have paid no attention the ioc flag. Following scenario
occurred occasionally (we could observe error messages in the kernel
log):

There is a TD consisting of multiple TRBs on the transfer ring. We
retrieve the TRB one by one, forward the request to the device, handle
the DMA (DMA would come first in case of OUT transfers), and always send
a transfer event (for each TRB). However, the driver expects us to only
send a transfer event for the last TRB, which it indicates by setting
`chain == true` and `ioc == false` on all but the last TRB, and `chain
== false` and `ioc == true` on the last TRB.

This commit adds a check for the ioc flag and skips the transfer event
sending when the driver expects no event.

While this small change gets rid of our error messages in the kernel
log, we would still fail in some scenarios. For example, when we process
a multi-TRB TD and one of the earlier TRBs is a short transfer (and
Interrupt-on-Short, ISP, is set), we need to send a transfer event for
this TRB and skip processing all remaining TRBs of the same TD. We have
not experienced such cases yet, and refining the implementation is not a
priority right now, but we should be aware.