nusb: circumvent timeout in transfer IN handling
nusb supports two ways of getting the results of a submitted transfer
request: async .await or synchronous wrappers with timeouts. Timeouts
are not ideal because we might have to wait an arbitrary time until an
interrupt IN endpoint sends data.

We should consider converting more to async Rust over time, for now I'm
choosing the simple fix of taking the longest timeout possible.

What if a bulk endpoint times out though? It is the driver's
responsibility to take action if it thinks that a request took too long,
e.g., by stopping and reseting the endpoint. We do not support that
commands as of now, but that approach will be the correct one in the
future, not ourselves choosing hopefully appropiate timeouts.