nusb: make sure wake up fails when worker thread dies
Previously, we just did "fire and forget". However, currently our
assumption is that endpoint worker threads live for the duration of the
usbvfiod runtime. When an endpoint panics (because of unexpected
interactions with the real endpoint), we would like the main thread to
panic as well as continuation without an endpoint worker does not make
sense.