flake: Use dried-nix-flakes to manage the input/outputs
So, what even is Dried Nix Flakes?

 - https://github.com/cyberus-technology/dried-nix-flakes

It's a pure-nix, and opinionated, tooling that aims to strip away the
repetitions found in a Flake.

Namely, it removes\* all mentions of the `system` from the Flake.

\*Actually not removed, but abstracted away.

With the repetitious mentions of `system` stripped-away, the Flake is
*DRY*er, as in "Don't Repeat Yourself"-er.

This fixes one of the major flaws with Flakes, where the CLI and Flake
expressions both end-up having to replace the `builtins.currentSystem`
value, which was deemed “impure”. The Dried Nix Flakes Flake takes the
system outputs that Nixpkgs exposes by default, and automatically
collapses and expands the different standard Flakes attribute sets, such
that the need to use the `${system}` is removed, for the expressions.

With this change, the Flake is 1-to-1 equivalent for the systems that
were exposed in the parent commit.

The only “non-trivial” change was replacing `flake-utils.lib.mkApp`.
Since the `system` gets expanded by the Dried Nix Flakes semantics,
there is no need for an equivalent helper function.

This fixes https://github.com/cyberus-technology/usbvfiod/issues/108.