nix/tests: Prevent smoke test CI from running for aarch64-linux on Hercules CI
The interaction with GitHub CI status is weird.

The main issue seems to be the long time it takes to build the test,
which reaches a timeout. This timeout (1 hour) may be conflicting with
another timeout within the merge queue.

It's also unclear if the failure status was solely because of the
timeout, or if GitHub somehow knows about the ignored failure and made
things harder to understand.

This was checked with:

 - `nix flake check`
 - `nix --system aarch64-linux flake check --no-build`
 - and:

```
 $ nix-instantiate --eval --strict --expr '{ flake = builtins.getFlake (toString ./.); }' --attr flake.checks.x86_64-linux.integration-smoke.buildDependenciesOnly
false

 $ nix-instantiate --eval --strict --expr '{ flake = builtins.getFlake (toString ./.); }' --attr flake.checks.aarch64-linux.integration-smoke.buildDependenciesOnly
true
```